pgsql/log: rename back to logger, makes more sense
